requirements.txt
streamlit>=1.28.0
streamlit-folium>=0.15.0
folium>=0.15.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
xgboost>=1.7.0
joblib>=1.3.0
scikit-learn>=1.3.0
earthengine-api>=0.1.380
geemap>=0.30.0
geopandas>=0.14.0
pyyaml>=6.0
